mmp2/irq: add ranges property to /interrupt-controller
The "regs" property of the "mrvl,mmp2-mux-intc" devices are offsets from
intc's base, not addresses on the parent bus. This shouldn't have happened,
but it had, and the bindings can't be changed.

To add insult to injury, Linux's mmp2 irqchip driver gets the ranges
with of_address_to_resource() as if it were bus addresses. To make it
happy, let's pretend they actually are...

TODO: simple-bus is probably not needed...