mmp2/uart: specify reg-shift  property
This makes the Linux's 8250_of driver happy. It is broken: there are two
more drivers (yes...) that claim mrvl,mmp-uart compatibility and both assume
register shift of 2 bits. The third one should do the same.

However, we can't really tell or decide which driver is going to be
bound, so let's just add the property before 8250_of gets fixed. It
doesn't hurt to be a little bit more explicit anyway.