arm/linux: ignore the size from the zImage header
We already got the file size and that one is good enough.

Actually, it's one better: with CONFIG_ARM_APPENDED_DTB you can slap on
the DTB to the end of the image file, and still have the kernel see it.
If we were only considering the zImage size, we'd be blind to that.