usb2/keyboard: fix the "\" key
I'm unable to type a Backslash. None of the keyboards I have seems to be
able to produce the 0x31 scan code. All of them have a key that produces
scan code 0x64 just to the right of the left Shift key, labeled as follows:

  | *
  \ &

That seems to be the "Keyboard Non-US \ and |" key. Oddly enough, Linux
with US English layout considers this a "<" and ">" (shifted) key.
There's another pair of "<" and ">" keys at the other end of the row.

There's another Backslash key with code 0x32 on keyboard with Czech layout,
just above the right Shift key, left of Return:

  | '
  \ ¨

That is somewhat closer to where it would it be on an US keyboard. The
Return key on these keyboards spans two rows, occupying also the space
where the US "\" key would be. On a keyboard with English keys + Euro
key it is labeled differently:

 ~
 #

The HID boot descriptor seems to suggest that indeed this should be a
"Keyboard Non-US # and ~" key. However, Linux with US English layout
uses this as a Backslash/Pipe key.

I'm labeling both keys as Backslash/Pipe. The 0x32 one matches the
labels on all the keyboards around here (Czech and English International?).
The 0x64 one matches the Czech layout, and, more importantly, what an
user would actually expect because that's how the key usually works in
the OS. Both keys were undefined until now so I guess I'm unlikely to
ruin things for anyone.