pc/olpc: link /camera to /image-sensor
Draw a device graph edge between the two, so that the driver will be
able to discover how are they connected. We need to do this after the
PCI device is scanned, because the CAFE function drop-in will create the
necessary nodes, but the FCode does not have a provision for drawing
device graphs.