mmp2/uart: make the MMP UART compatible with intel,xscale-uart
The UART is XScale compatible and the more specific compatible string
shouldn't really have been necessary.

There are three drivers that bind to mrvl,mmp-uart in Linux, for
legacy reasons. The 8250_of driver should be preferrably used and the
other two sometimes break.

We need to keep the more specific compatible string for OLPC builds that
lack the 8250_of driver, but prefer the XScale one.