olpc: turn on olpc-compat? if alternate? is inspected on boot
The OLPC OS on XO-4 (be it in secured mode, or on unsecured laptop booted
from an USB flash stick) executes the alternate? word to find out if it
shall boot the Android kernel or whatever it is.

While probably quite half-hearded, that seems to be the only way to tell if
we're running OLPC OS, that probably has a 3.5-based kernel that needs a
compatibility device tree.

The other distros that are potentially going to ship newer kernels have
no need to call alternate? in their boot paths.