mmp3/hdmi: avoid accidentally writing to some low addresses
It is not clear what is the point the c! here. What's on the stack at
that point is the index into he packet, but a memory address and
executing in just clobbers something in low memory addresses.

Even if it was and address and not an index, writing back what we just
read does not seem to make much sense either -- if I'm reading this
right.

Let's just drop index to conform with the stack diagram. I don't have
equipment picky enough about AVI info frame to test this; ordinary HDMI
and DVI capable computer monitors don't seem to care.