olpc/lcd: enable HDMI on startup if LCD0 resolution is supported
This does nothing on the XO, because the resolution of the panel on LCD0
is unlike anything mmp3/hdmi knows about.

On the Dell Ariel board, however, there's no internal panel. There's a
DVI-I port driven by LCD0 via an external encoder and a DVI-D port
driven via LCD1's HDMI encoder. We want to bring both up with some safe
resolution, such as 1024x768, and this looks like a reasonable place to
do just that.