olpc/sdhci: add vmmc-supply for eMMC on XO-4
Unlike on XO-1.75, on XO-4, power to eMMC is controlled by a GPIO and the
mainline kernel doesn't follow the power-gpios property. Instead, it
needs an instance of a fixed regulator, defined by
regulator/fixed-regulator.txt.

Create one and bind it to MMC controller eMMC card is connected to
following the regulator/regulator.txt and mmc/mmc.txt bindings.